chore: move types out of __init__ file
move types in /service_layer/__init__.py to types.py
