+++ src/adapters/tracking_repository.py
@@ -4,7 +4,7 @@ from typing import Generic, List, Optional, Set
 from src.adapters.repository import AbstractRepository
 from src.core.types import UUID
 from src.domain.model import Task
-from src.service_layer import RepoType
+from src.service_layer.types import RepoType
 
 
 class TrackingRepository(AbstractRepository, Generic[RepoType]):
