+++ src/service_layer/queue/broker.py
@@ -2,9 +2,9 @@ from typing import Set
 
 from src.domain.commands import Command
 from src.domain.events import Event
-from src.service_layer import Message
 from src.service_layer.queue.command_queue import CommandQueue
 from src.service_layer.queue.event_queue import EventQueue
+from src.service_layer.types import Message
 
 
 class MessageBroker:
