+++ src/service_layer/types.py
@@ -1,0 +1,11 @@
+from typing import TypeVar
+
+from src.adapters.repository import AbstractRepository
+from src.domain.commands import Command
+from src.domain.events import Event
+
+type Message = Event | Command
+RepoType = TypeVar(
+    "RepoType", bound=AbstractRepository, default=AbstractRepository, covariant=True
+)
+CallbackType = TypeVar("CallbackType", bound=Message)
